Agents/server/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
fastapi==0.121.1
filetype==1.2.0
fsspec==2024.12.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-auth==2.43.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
json_repair==0.39.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.79
langchain-google-genai==3.0.2
langchain-neo4j==0.4.0
langchain-text-splitters==0.3.11
langsmith==0.4.42
neo4j==5.28.2
neo4j-graphrag==1.6.1
orjson==3.11.4
packaging==25.0
proto-plus==1.26.1
protobuf==6.33.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.4
pydantic_core==2.41.5
pypdf==5.9.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
tenacity==9.1.2
types-PyYAML==6.0.12.20250915
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
zstandard==0.25.0
